refactor: remove comentarios, separa modulos do Autocomplete e isola responsabilidades
