+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/FiltersBar/modules/common/FiltersSelect/modules/AutoComplete/modules/MultiSelect/index.tsx
@@ -1,0 +1,56 @@
+import type { Dispatch, JSX, SetStateAction } from "react";
+import Autocomplete from "@mui/material/Autocomplete";
+import { sxMultipleSelect } from "./consts";
+
+import type { AppliedFilters } from "@/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/common/SharedAppliedFilters";
+import type {
+    HtmlSelectOption,
+    SelectConfig,
+} from "@features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/FiltersBar/modules/common/SelectConfig";
+import { slotProps } from "../common/slotProps";
+import { renderInput } from "../common/renderInput";
+import { renderOptionMultiSelect, renderTagsMultiSelect } from "./logic";
+
+type AutoCompleteProps<TAppliedFilters extends AppliedFilters> = {
+    value: Array<HtmlSelectOption>;
+    // TODO: rever se o componente precisa receber o select inteiro ou apenas o id e as opções.
+    select: SelectConfig;
+    setSelectedValues: Dispatch<SetStateAction<TAppliedFilters>>;
+    selectedValues: TAppliedFilters;
+};
+
+export const AutoCompleteMultiSelect = <
+    TAppliedFilters extends AppliedFilters,
+>({
+    value,
+    select,
+    setSelectedValues,
+    selectedValues,
+}: AutoCompleteProps<TAppliedFilters>): JSX.Element => {
+    return (
+        <Autocomplete
+            value={value}
+            key={select.id}
+            onChange={(_event, newValue) => {
+                setSelectedValues((prevState) => ({
+                    ...prevState,
+                    [select.id]: newValue.map((item) => item.value),
+                }));
+            }}
+            multiple={true}
+            options={select.options}
+            limitTags={1}
+            sx={sxMultipleSelect(
+                selectedValues[
+                    select.id as keyof TAppliedFilters
+                ] as Array<string>
+            )}
+            disableCloseOnSelect
+            renderInput={renderInput(select.label)}
+            renderOption={renderOptionMultiSelect}
+            renderTags={renderTagsMultiSelect}
+            slotProps={slotProps}
+            style={{ width: select.width }}
+        />
+    );
+};
