+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/FiltersBar/modules/common/FiltersSelect/modules/AutoComplete/modules/common/renderInput.tsx
@@ -1,0 +1,9 @@
+import type { AutocompleteRenderInputParams } from "@mui/material";
+import { TextField } from "@mui/material";
+import type { JSX } from "react";
+
+export const renderInput = (label: string) => {
+    return (params: AutocompleteRenderInputParams): JSX.Element => (
+        <TextField {...params} label={label} />
+    );
+};
